core: split external script unit test and main code in diff file
split the external script main code and unit test in different file.

Closes: https://github.com/rook/rook/issues/9588
